+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 
+## 1.0.0.0
+
+*   Add the functions `relaxOpenUnion` and `relaxUnion` for being able to add
+    additional types to an `OpenUnion`. [#3]
+*   Add the functions `openUnionRemove`, `unionRemove`, `openUnionHandle`, and
+    `unionHandle` for being able to pull individual types out of an
+    `OpenUnion`. [#3]
+
 ## 0.1.0.0
 
 *   Initial release.